estimation/requirements.txt
numpy
opencv-python
loguru
scikit-image
scikit-learn
tqdm
torchvision>=0.10.0
Pillow
thop
ninja
tabulate
lap
motmetrics
filterpy
h5py
matplotlib
scipy
prettytable
easydict
tensorboard
pyyaml
yacs
termcolor
gdown
# onnx==1.8.1
# onnxruntime==1.8.0
# onnx-simplifier==0.3.5


